api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
azure-eventhub>=5.12.0
pymongo
redis>=5.0